+++ pypeit/images/pypeitimage.py
@@ -21,6 +21,7 @@ from pypeit import utils
 from pypeit import masterframe
 
 
+# TODO This docs need to indicate the types of things. Are crmask fullmask etc. boolean or ints?
 class PypeItImage(datamodel.DataContainer):
     r"""
     Container class for processed ``PypeIt`` images and associated data.
@@ -33,6 +34,10 @@ class PypeItImage(datamodel.DataContainer):
 
     Note that all arguments for instantiation are optional, meaning that an
     empty object can be created.
+
+    The datamodel attributes are:
+
+    .. include:: ../include/class_datamodel_pypeitimage.rst
 
     Args:
         image (`numpy.ndarray`_, optional):
@@ -53,7 +58,6 @@ class PypeItImage(datamodel.DataContainer):
             Image count scaling applied (e.g., 1/flat-field)
         bpm (`numpy.ndarray`_, optional):
             Bad pixel mask
-        # TODO This docs need to indicate the types of things. Are crmask fullmask etc. boolean or ints?
         crmask (`numpy.ndarray`_, optional):
             CR mask image
         fullmask (`numpy.ndarray`_, optional):
@@ -121,7 +125,9 @@ class PypeItImage(datamodel.DataContainer):
                  'crmask': dict(otype=np.ndarray, atype=np.bool_, descr='CR mask image'),
                  'fullmask': dict(otype=np.ndarray, atype=np.integer, descr='Full image bitmask'),
                  'detector': dict(otype=(DetectorContainer, Mosaic),
-                                  descr='The detector or mosaic parameters'),
+                                  descr='The detector (see :class:`~pypeit.images.detector_container.DetectorContainer`) '
+                                        'or mosaic (see :class:`~pypeit.images.mosaic.Mosaic`) '
+                                        'parameters'),
                  'PYP_SPEC': dict(otype=str, descr='PypeIt spectrograph name'),
                  'units': dict(otype=str, descr='(Unscaled) Pixel units (e- or ADU)'),
                  # TODO: Consider forcing exptime to be a float.