+++ pypeit/pypeitsetup.py
@@ -220,11 +220,7 @@ class PypeItSetup:
             os.makedirs(output_path)
         
         # Grab the list of files
-        data_files = setup.files_from_extension()
-        dfname = os.path.join(root, '*{0}*'.format(extension)) \
-                    if os.path.isdir(root) else '{0}*{1}*'.format(root, extension)
-        data_files = glob.glob(dfname)
-        data_files.sort()
+        data_files = setup.files_from_extension(root, extension=extension)
 
         # Instantiate
         return cls.from_rawfiles(data_files, spectrograph)
@@ -384,6 +380,9 @@ class PypeItSetup:
         # configuration-defining metadata
         if clean_config:
             self.fitstbl.clean_configurations()
+            if len(self.fitstbl) == 0:
+                msgs.error('Cleaning the configurations removed all the files!  Rerun '
+                           'pypeit_setup with the --keep_bad_frames option.')
 
         # Determine the type of each frame.
         self.get_frame_types(flag_unknown=setup_only or calibration_check)
@@ -479,6 +478,7 @@ class PypeItSetup:
 
         if obslog:
             log_file = pypeit_file.replace('.pypeit', '.obslog')
+            log_file = os.path.join(_output_path, os.path.split(log_file)[1])
             header = ['Auto-generated PypeIt Observing Log',
                       '{0}'.format(time.strftime("%a %d %b %Y %H:%M:%S", time.localtime()))]
             self.fitstbl.write(output=log_file, columns='pypeit', sort_col='mjd',