+++ pypeit/scripts/__init__.py
@@ -4,6 +4,7 @@ from pypeit.scripts import scriptbase
 
 # The import of all the script modules here is what enables the dynamic
 # compiling of all the available scripts below
+from pypeit.scripts import cache_github_data
 from pypeit.scripts import chk_alignments
 from pypeit.scripts import chk_edges
 from pypeit.scripts import chk_flats
@@ -29,9 +30,7 @@ from pypeit.scripts import obslog
 from pypeit.scripts import parse_calib_id
 from pypeit.scripts import parse_slits
 from pypeit.scripts import qa_html
-from pypeit.scripts import ql_keck_deimos
-from pypeit.scripts import ql_keck_nires
-from pypeit.scripts import ql_mos
+from pypeit.scripts import ql
 from pypeit.scripts import ql_multislit
 from pypeit.scripts import run_pypeit
 from pypeit.scripts import sensfunc