+++ pypeit/spectrographs/gtc_osiris.py
@@ -22,6 +22,7 @@ class GTCOSIRISSpectrograph(spectrograph.Spectrograph):
     name = 'gtc_osiris'
     telescope = telescopes.GTCTelescopePar()
     camera = 'OSIRIS'
+    url = 'http://www.gtc.iac.es/instruments/osiris/'
     header_name = 'OSIRIS'
     supported = True
     comment = 'See :doc:`gtc_osiris`'
@@ -127,9 +128,15 @@ class GTCOSIRISSpectrograph(spectrograph.Spectrograph):
         par['calibrations']['tiltframe']['process']['combine'] = 'mean'
         par['calibrations']['tiltframe']['process']['subtract_continuum'] = True
 
+        # Increase the wave tilts order, given the longish slit
+        par['calibrations']['tilts']['spat_order'] = 5
+        par['calibrations']['tilts']['spec_order'] = 5
+
         #Only extract one object per standard frame
         par['reduce']['findobj']['maxnumber_std']=1
 
+        # Turn off the 2D fit - this seems to be giving bad results for OSIRIS
+        par['reduce']['skysub']['no_poly'] = True
         return par
 
     def init_meta(self):
@@ -335,6 +342,8 @@ class GTCOSIRISSpectrograph(spectrograph.Spectrograph):
             # par['calibrations']['wavelengths']['disp'] = 1.36
             par['calibrations']['wavelengths']['lamps'] = ['ArI,XeI,NeI']
             par['calibrations']['wavelengths']['reid_arxiv'] = 'gtc_osiris_R2500I.fits'
+            par['sensfunc']['algorithm'] = 'IR'
+            par['sensfunc']['IR']['telgridfile'] = "TelFit_MaunaKea_3100_26100_R20000.fits"
         else:
             msgs.warn('gtc_osiris.py: template arc missing for this grism! Trying holy-grail...')
             par['calibrations']['wavelengths']['method'] = 'holy-grail'