+++ pypeit/spectrographs/spectrograph.py
@@ -27,6 +27,7 @@ provide instrument-specific:
 """
 
 from abc import ABCMeta
+import os
 
 from IPython import embed
 
@@ -97,6 +98,11 @@ class Spectrograph:
     This is used by specdb, so use that naming convention
     """
 
+    url = None
+    """
+    Reference url
+    """
+
     header_name = None
     """
     Name of the spectrograph camera or arm from the Header.
@@ -133,7 +139,12 @@ class Spectrograph:
     Metadata model that is generic to all spectrographs.
     """
 
-    def __init__(self, quicklook=False):
+    allowed_extensions = None
+    """
+    Defines the allowed extensions for the input fits files.
+    """
+
+    def __init__(self, quicklook=True):
         self.dispname = None
         self.rawdatasec_img = None
         self.oscansec_img = None
@@ -206,6 +217,21 @@ class Spectrograph:
         """
 
         return par
+
+    def _check_extensions(self, filename):
+        """
+        Check if this filename has an allowed extension
+
+        Args:
+            filename (:obj:`str`):
+                Input raw fits filename
+        """
+        if self.allowed_extensions is not None:
+            if os.path.splitext(filename)[1] not in self.allowed_extensions:
+                msgs.error("The input filename:"+msgs.newline()+
+                           filename+msgs.newline()+
+                           f"has the wrong extension. The allowed extensions for {self.name} include:"+msgs.newline()+
+                           ",".join(self.allowed_extensions))
 
     def _check_telescope(self):
         """Check the derived class has properly defined the telescope."""
@@ -653,7 +679,8 @@ class Spectrograph:
         if 'instrument' in meta_tbl.keys():
             # Check that there is only one instrument
             #  This could fail if one mixes is much older calibs
-            instr_names = np.unique(meta_tbl['instrument'].data)
+            indx = meta_tbl['instrument'].data != None
+            instr_names = np.unique(meta_tbl['instrument'].data[indx])
             if len(instr_names) != 1:
                 msgs.warn(f"More than one instrument in your dataset! {instr_names} \n"+
                 f"Proceed with great caution...")
@@ -999,7 +1026,8 @@ class Spectrograph:
             pixel. Pixels unassociated with any amplifier are set to 0.  Shape
             is identical to ``raw_img``.
         """
-        # Open
+        # Check extension and then open
+        self._check_extensions(raw_file)
         hdu = io.fits_open(raw_file)
 
         # Validate the entered (list of) detector(s)
@@ -1403,6 +1431,7 @@ class Spectrograph:
         # Faster to open the whole file and then assign the headers,
         # particularly for gzipped files (e.g., DEIMOS)
         if isinstance(inp, str):
+            self._check_extensions(inp)
             try:
                 hdu = io.fits_open(inp)
             except: